+++ NEWS.md
@@ -1,5 +1,7 @@
 # cheapr (Development version)
 
+* New function `recycle` to help in recycling R objects to a common size.
+
 * The `set` functions that update by reference are now ALTREP aware and
 take a copy when the input is an ALTREP object.